fix:  seguiment progrés lectura local amb gun
