+++ components/FabricaCapitols/TileCapitol.js
@@ -1,28 +1,30 @@
 import Link from 'next/link';
-import {useState,useEffect} from 'react';
-import {ultimCapitol} from '../../utils/MyGun';
+import {useState,useEffect, useContext} from 'react';
+
+import {gunContext} from '../../pages/_app';
 
 function TileCapitol(props) {
 
+    const [llegit,setLlegit] = useState("is-dark");
+  
+    //Gun
+    const {gun,nomGraph} = useContext(gunContext);    
     const { titol, slug } = props.capitol;   
     const url = `${props.path}/${slug}`;
-    const [llegit,setLlegit] = useState("is-dark");
-
+    const doc = props.path.split('/')[2];   
+    const docKey = doc.split('-')[1];    
+    
     useEffect(() => {
       
-        ultimCapitol('get',url).then((res,err)=>{
+        gun.get(nomGraph).get(docKey).get(slug).get("llegit").once((res) => {
 
-            console.log(`res al component ${slug}`, res);
+            console.log(`resposta del get ${slug}`, res);
 
-            if(err){
-                console.error(err);
-            }     
-     
-            if(res._.put==true){
-                setLlegit('is-light');
-            }
+            if(res===true){
+                setLlegit("is-light");
+            }           
 
-        });      
+        },{wait: 0})   
 
     }, [])
 
@@ -37,13 +39,11 @@ function TileCapitol(props) {
                 </p>        
             </div>
             <footer className="card-footer">
-                <p className="card-footer-item">
-                
+                <p className="card-footer-item">                
                     <Link href={url}>
                     <a className={`button ${llegit} is-fullwidth`}>{llegit==="is-dark"?"Llegir":"Tornar a llegir"}</a>
                     </Link>               
-                </p>
-                
+                </p>                
             </footer>
         </div>
